Resolve #572 -- Buffs Rework (#954)
* Better Syntax for Buff Variables
* Transition Buff Removal to Deactivation
* Buffs show their correct time on the hud.
* Improved efficiency regarding usage of buff slots.
* Resolve Particle Removal
* Adds OnRemoved  override to Particle class and moves packet notification to it.
* Resolve Natural Removals
* Implement Abstract Stackable Class
* Use LINQ for Buff foreach statements in ObjAiBase
* Implemented reverse for loop in ObjectManager for natural Buff removals.
* Cleaned ObjAiBase.HasBuff function.
* Added exception call to Buff class for when a Stackable Buff is instantiated with a MaxStacks less than 2.

+++ GameServerLib/Items/Shop.cs
@@ -31,7 +31,7 @@ namespace LeagueSandbox.GameServer.Items
             _owner.Stats.Gold += sellPrice;
             
             i.DecrementStackCount();
-            RemoveItem(i, slotId, i.StackCount);
+            RemoveItem(i, slotId, (byte)i.StackCount);
             return true;
         }
 
@@ -59,7 +59,7 @@ namespace LeagueSandbox.GameServer.Items
                 foreach (var item in ownedItems)
                 {
                     item.DecrementStackCount();
-                    RemoveItem(item, inventory.GetItemSlot(item), item.StackCount);
+                    RemoveItem(item, inventory.GetItemSlot(item), (byte)item.StackCount);
                 }
 
                 AddItem(itemTemplate);
